code-andre/requirements.txt
fastapi
uvicorn[standard]
chromadb
boto3
python-dotenv
pydantic
requests
beautifulsoup4
docling